Make `@cache.invalidate()` idempotent
Problem: `invalidate()` raises `KeyError` if the key is missing.

That leads to a nasty behavior: code works under unit tests,
but raises the error in production (because of the `timeout`).

An especially nasty case is when some concurrency is involved,
where two writer threads execute some write operation simultaneously,
and then one thread wins the race, while another gets `KeyError`.

In practice, it is almost impossible to write any deterministic logic
that will do something like:
  "I know I mutated this specific thing, so invalidate the old value"

so I always have to do this:

    try:
        function_decorated_with_cache.invalidate(key)
    except KeyError:
        pass

And I guess everyone who faced this issue has to do the same,
so it looks to me more a bug than a feature.

So here I decided to fix it.
Now `invalidate()` just doesn't raise `KeyError` if the key is
missing.

Also, I've added a separate test dedicated to invalidation,
where I also added `invalidate_all()` (that wasn't covered by tests
previously).